--- src/custom-scripts/examples/demo.lua
@@ -1,4 +1,0 @@
-function foo(arg1)
-    if arg1 == '1' then return "Wizard Class" end
-    if arg1 == '2' then return "Warrior Class" end
-end
