--- src/custom-scripts/examples/dynamic_string.lua
@@ -1,4 +1,0 @@
-function foo(arg1)
-    if arg1 == '1' then return "You can't taste the rock" end
-    if arg1 == '2' then return "YOU CAN'T TASTE THE ROCK" end
-end
