+++ src/custom-scripts/examples/lua/demo.lua
@@ -1,0 +1,4 @@
+function foo(arg1)
+    if arg1 == '1' then return "../../../../src/custom-scripts/examples/wdl/demo-wizard.wdl" end
+    if arg1 == '2' then return "../../../../src/custom-scripts/examples/wdl/demo-warrior.wdl" end
+end
